changed the includes of the files so we don't have a centrilized header
