+++ src/ClientInfo.cpp
@@ -1,4 +1,4 @@
-#include "../inc/Headers.hpp"
+#include "ClientInfo.hpp"
 
 ClientInfo::ClientInfo()
 {
