+++ src/Node.cpp
@@ -1,4 +1,4 @@
-#include "../inc/Headers.hpp"
+#include "Node.hpp"
 
 Node::Node()
 {
